pyproject.toml
[project]
name = "trading-multi-agent"
version = "1.0.0"
description = "Production-grade Multi-Agent Trading System with LangGraph orchestration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "tortoise-orm[asyncpg]",
    "aerich",
    "asyncpg",
    "psycopg2-binary",
    "pyzmq",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langchain-anthropic",
    "langgraph",
    "langchain-chroma",
    "chromadb",
    "openai",
    "anthropic",
    "sentence-transformers",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "httpx",
    "aiofiles",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[project.scripts]
trading-api = "uvicorn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
